# Documentacion/openapi.yaml
openapi: 3.0.3
info:
  title: Plants App API
  description: |
    API REST del TFG "Plants App" — aplicación móvil de gestión de plantas con
    funcionalidades de comunidad, chat en tiempo real, recordatorios y datos climáticos.
    Arquitectura: Clean Architecture (Node.js + Express + MongoDB + Socket.IO).
  version: 1.0.0
  contact:
    name: TFG — 2 DAM
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo local

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTES COMUNES
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido tras login o registro.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
          example: Token inválido o expirado.
      required: [status, code, message]

    # ── Paginación ────────────────────────────────────────────────────────────
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        hasMore:
          type: boolean
          example: true

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    Post:
      type: object
      properties:
        id:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d3
        userId:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d1
        authorName:
          type: string
          example: Ana García
        authorPhoto:
          type: string
          nullable: true
        content:
          type: string
          maxLength: 1000
          example: Mi monstera ha echado una nueva hoja!
        image:
          type: string
          nullable: true
        likesCount:
          type: integer
          example: 42
        commentsCount:
          type: integer
          example: 5
        isLikedByMe:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePostRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
          example: Mi monstera ha echado una nueva hoja!
        imageUrl:
          type: string
          nullable: true
        plantId:
          type: string
          nullable: true
    Comment:
      type: object
      properties:
        id:
          type: string
        postId:
          type: string
          description: ID del post al que pertenece el comentario.
        userId:
          type: string
        authorName:
          type: string
        authorPhoto:
          type: string
          nullable: true
        content:
          type: string
          maxLength: 500
        createdAt:
          type: string
          format: date-time
    CreateCommentRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 500

    # ── Chat ──────────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: No autorizado — token ausente o inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Prohibido — rol insuficiente.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Error de validación en el cuerpo de la petición.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /community:
    get:
      tags: [Community]
      summary: Obtener feed de la comunidad
      description: Devuelve posts paginados ordenados por fecha descendente.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Feed paginado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Community]
      summary: Crear post
      description: Publica un nuevo post en la comunidad.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /community/mine:
    get:
      tags: [Community]
      summary: Obtener posts propios (perfil propio)
      description: |
        Variante del feed que devuelve únicamente los posts del usuario
        autenticado (modo "Mi perfil"). Implementado en F3-042 / decisión #59
        sustituyendo el filtrado local del antiguo `UserProfileViewModel`.
        Backend ejecuta `IGetFeedUseCase.execute(userId, page, limit,
        authorId=userId)` con `mine=true` interno.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Posts propios paginados.
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /community/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ID del post.
    get:
      tags: [Community]
      summary: Obtener post por ID
      responses:
        '200':
          description: Detalle del post.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Community]
      summary: Eliminar post propio
      description: Soft-delete del post. Solo el autor puede eliminarlo (ownership validation).
      responses:
        '204':
          description: Post eliminado.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /community/{id}/like:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ID del post.
    post:
      tags: [Community]
      summary: Dar like a un post
      description: Registra un like del usuario autenticado. Devuelve 409 si el usuario ya había dado like previamente.
      responses:
        '204':
          description: Like registrado correctamente.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: El usuario ya ha dado like a este post (LIKE_ALREADY_EXISTS).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Community]
      summary: Quitar like de un post
      responses:
        '204':
          description: Like eliminado.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /community/{id}/comments:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ID del post.
    get:
      tags: [Community]
      summary: Obtener comentarios de un post
      responses:
        '200':
          description: Lista de comentarios.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Community]
      summary: Comentar un post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comentario creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /community/{id}/comments/{commentId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ID del post.
      - name: commentId
        in: path
        required: true
        schema:
          type: string
        description: ID del comentario.
    delete:
      tags: [Community]
      summary: Eliminar comentario propio
      description: |
        Soft-delete del comentario. Solo el autor puede eliminarlo (ownership validation).
        Decrementa commentsCount del post padre. Broadcast post:updated via socket.
      responses:
        '204':
          description: Comentario eliminado.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════════════════════════════════════════
  # CHAT
  # ════════════════════════════════════════════════════════════════════════════
...
  /admin/posts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ID del post.
    delete:
      tags: [Admin]
      summary: Eliminar post por moderación (admin)
      description: |
        Soft-delete de un post por motivos de moderación.
        Envía notificación in-app al autor indicando que su contenido fue eliminado.
        Solo administradores.
      responses:
        '204':
          description: Post eliminado por moderación.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...